requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
sqlparse==0.4.3
whitenoise==6.3.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7

